+++ CHANGELOG.md
@@ -334,6 +334,10 @@ Current
 
 ### Deprecated:
 
+- [Remove `PhysicalTable::getTableName` to use `getName` instead](https://github.com/yahoo/fili/pull/263)
+    * Having more than 1 method for the same concept (ie. what's the name of this physical table) was confusing and not
+      very useful.
+
 - [Remove `PhysicalTableDictionary` dependency from `SegmentIntervalHashIdGenerator`](https://github.com/yahoo/fili/pull/263)
     * Constructors taking the dictionary have been deprecated, since it is not used any more